+++ scapi/__init__.py
@@ -1,30 +1,26 @@
 # Created by kakeruzoku / https://github.com/kakeruzoku/scapi
 # Special Thanks: Timmccool / https://github.com/TimMcCool/scratchattach
 
-__version__ = "0.0.3"
+__version__ = "0.1.0"
 
 from .others.common import (
     create_ClientSession,
     Response,
-    Requests,
+    ClientSession,
     api_iterative as _api_iterative,
-    split_int as _split_int,
-    split as _split,
-    to_dt as _to_dt,
+    split_int,
+    split,
+    to_dt,
     empty_project_json,
     BIG
 )
-from .others.error import *
-del TYPE_CHECKING
+from .others import error as exception
+del exception.TYPE_CHECKING
 from .others.other_api import (
     get_csrf_token_sync
 )
 from .sites.base import (
     _BaseSiteAPI,
-    get_object as _get_object,
-    get_object_iterator as _get_object_iterator,
-    get_comment_iterator as _get_comment_iterator,
-    get_count as _get_count
 )
 from .sites.comment import (
     CommentData,